Extract common renderer setup
